chore: move unit tests to subdirectory
