+++ package.json
@@ -32,7 +32,7 @@
     "preset": "@theholocron/jest-preset",
     "testEnvironment": "jest-environment-node",
     "testPathIgnorePatterns": [
-      "<rootDir>/test/mocks.js",
+      "<rootDir>/test/unit/mocks.js",
       "<rootDir>/node_modules/"
     ],
     "transform": {}
@@ -49,7 +49,7 @@
     "dist": "echo 'No dist task specified' && exit 1",
     "lint": "eslint .",
     "release": "semantic-release",
-    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js test --passWithNoTests"
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js test/unit --passWithNoTests"
   },
   "devDependencies": {
     "@theholocron/commitlint-config": "^0.1.11",
